Add caption translation generator.
This takes in a .srt caption file (of multiple languages) and outputs to
another .srt file, one for each desired output language.

This script uses google translate api and requires a google cloud
account and a project to be set up with the translate api.
Refer to
https://codelabs.developers.google.com/codelabs/cloud-translation-python3#0
For instructions to how to set up your local machine for such
development, and download the google cloud client key info file to your computer
locally.