Added experimental two-part translation process:
First, the captions are translated to Korean and English.
Then they are reviewed, which is typically a lot quicker than the other
two.
Then you translate the other two based on the Korean and English
translations, to take advantage of grammatical similarities between the
two languages.

Additionally renamed a subtitle generating script in process_sequence.py